ready to release add proc_byp_t

+++ src/ap_file.c
@@ -1,11 +1,10 @@
-//
-//  ap_file.c
-//  ap_file_system
-//
-//  Created by HU XUKAI on 14/11/12.
-//  Copyright (c) 2014年 HU XUKAI.<goingonhxk@gmail.com>
-//
-
+/*
+ *   Copyright (c) 2015, HU XUKAI
+ *
+ *   This source code is released for free distribution under the terms of the
+ *   GNU General Public License.
+ *
+ */
 #include <stdio.h>
 #include <string.h>
 #include <errno.h>
@@ -13,27 +12,18 @@
 #include <ap_fs.h>
 #include <ap_pthread.h>
 #include <bag.h>
-
-/*每当一个独立线程调用ap_open时都会产生新的file结构
- *以后能使用hash使得同一个文件对应一个fd（相对于不同线程来说）
- */
+#include <envelop.h>
 
 static inline int get_mask_from_oflag(int flag)
 {
-    int mask = 0;
-    switch (flag) {
-        case O_RDONLY:
-            mask = MAY_READ;
-            break;
-        case O_WRONLY:
-            mask = MAY_WRITE;
-            break;
-        case O_RDWR:
-            mask = MAY_WRITE | MAY_READ;
-            break;
-        default:
-            return -1;
-    }
+    int mask = MAY_READ;
+    if (flag & O_RDWR)
+        mask = MAY_READ | MAY_WRITE;
+    else if (flag & O_RDONLY)
+        mask = MAY_READ;
+    else if (flag & O_WRONLY)
+        mask = MAY_WRITE;
+    
     return mask;
 }
 
@@ -48,10 +38,6 @@ int ap_open(const char *path, int flags)
     }
     
     int mask = get_mask_from_oflag(flags);
-    if (mask == -1) {
-        errno = EINVAL;
-        return -1;
-    }
     
     SHOW_TRASH_BAG;
     struct ap_file *file;
@@ -60,7 +46,7 @@ int ap_open(const char *path, int flags)
     TRASH_BAG_PUSH(&final_indc->indic_bag);
     ap_fpthr = pthread_getspecific(file_thread_key);
     if (ap_fpthr == NULL) {
-        fprintf(stderr, "ap_thread didn't find\n");
+        ap_err("ap_thread didn't find\n");
         exit(1);
     }
     
@@ -85,11 +71,12 @@ int ap_open(const char *path, int flags)
     
     if ((ap_vfs_permission(final_indc, mask)) != 0){
         errno = EACCES;
-        return -1;
+        B_return(-1);
     }
     
-    
-    if (final_indc->cur_inode->f_ops->open != NULL) {
+    file->f_ops = final_indc->cur_inode->f_ops;
+
+    if (file->f_ops != NULL && file->f_ops->open != NULL) {
         int open_s;
         open_s = final_indc->cur_inode->f_ops->open(file, final_indc->cur_inode, flags);
         if (open_s == -1) {
@@ -97,7 +84,6 @@ int ap_open(const char *path, int flags)
         }
     }
     
-    file->f_ops = final_indc->cur_inode->f_ops;
     file->relate_i = final_indc->cur_inode;
     ap_inode_get(final_indc->cur_inode);
     
@@ -130,49 +116,60 @@ static int __ap_mount(void *m_info, struct ap_file_system_type *fsyst, const cha
     
     par_indic = MALLOC_INODE_INDICATOR();
     mount_point = MALLOC_AP_INODE();
-    
     mount_point->is_mount_point = 1;
 
-    mount_inode = fsyst->get_initial_inode(fsyst,m_info);
-    if (mount_inode == NULL) {
-        return -1;
-    }
-    if (mount_inode->links == 0) {
-        mount_inode->links++;
-    }
-    mount_point->real_inode = mount_inode;
-    mount_inode->mount_inode = mount_point;
-    
     struct ap_file_pthread *ap_fpthr = pthread_getspecific(file_thread_key);
     if (ap_fpthr == NULL) {
-        fprintf(stderr, "ap_thread didn't find\n");
+        ap_err("ap_thread didn't find\n");
         exit(1);
     }
     
-    if((initial_indicator(path, par_indic, ap_fpthr)) == -1)
+    if((initial_indicator(path, par_indic, ap_fpthr)) == -1){
+        AP_INODE_FREE(mount_point);
+        AP_INODE_INICATOR_FREE(par_indic);
         return -1;
+    }
     
     get = walk_path(par_indic);
     
     if ((get == -1 && par_indic->slash_remain > 0) ||
         !par_indic->cur_inode->is_dir) {
         errno = ENOENT;
+        AP_INODE_FREE(mount_point);
         AP_INODE_INICATOR_FREE(par_indic);
         return -1;
     }
     
-    if (!get) {
-        parent = par_indic->cur_inode->parent;
-    }else{
-        parent = par_indic->cur_inode;
+    if (par_indic->cur_mtp->fsyst != NULL &&
+        !par_indic->cur_mtp->fsyst->is_allow_mount) {
+        errno = ESRCH;
+        AP_INODE_FREE(mount_point);
+        AP_INODE_INICATOR_FREE(par_indic);
     }
+    
+    if (!get)
+        parent = par_indic->cur_inode->parent;
+    else
+        parent = par_indic->cur_inode;
+    
+    mount_inode = fsyst->get_initial_inode(fsyst,m_info);
+    if (mount_inode == NULL) {
+        AP_INODE_FREE(mount_point);
+        AP_INODE_INICATOR_FREE(par_indic);
+        return -1;
+    }
+    
+    if (mount_inode->links == 0)
+        mount_inode->links++;
+    
+    mount_point->real_inode = mount_inode;
+    mount_inode->mount_inode = mount_point;
     
     pthread_mutex_lock(&parent->ch_lock);
     if (!get){
         mount_point->prev_mpoints = par_indic->cur_inode;
         list_del(&par_indic->cur_inode->child);
     }
-    
     list_add(&mount_point->child, &parent->children);
     mount_point->parent = parent;
     mount_point->links++;
@@ -188,9 +185,8 @@ static int __ap_mount(void *m_info, struct ap_file_system_type *fsyst, const cha
     
     if (mount_inode->name == NULL) {
         dir_name = strrchr(path, '/');
-        if (strcmp(path, "/")) {
+        if (strcmp(path, "/"))
             dir_name++;
-        }
         len = strlen(dir_name);
         mount_inode->name = Malloc_z(len + 1);
         strncpy(mount_inode->name, dir_name, len);
@@ -217,7 +213,7 @@ extern int ap_mount2(char *file_system, const char *path)
     
     struct ap_file_system_type *fsyst = find_filesystem(file_system);
     if (fsyst == NULL || fsyst->get_mount_info == NULL) {
-        fprintf(stderr, "file_system didn't find\n");
+        ap_err("file_system didn't find\n");
         errno = EINVAL;
         return -1;
     }
@@ -238,7 +234,7 @@ int ap_mount(void *m_info, char *file_system, const char *path)
 
     struct ap_file_system_type *fsyst = find_filesystem(file_system);
     if (fsyst == NULL) {
-        fprintf(stderr, "file_system didn't find\n");
+        ap_err("file_system didn't find\n");
         errno = EINVAL;
         return -1;
     }
@@ -260,7 +256,7 @@ int ap_unmount(const char *path)
     struct ap_inode *op_inode, *mount_p;
     struct ap_file_pthread *ap_fpthr = pthread_getspecific(file_thread_key);
     if (ap_fpthr == NULL) {
-        perror("ap_thread didn't find\n");
+        ap_err("ap_thread didn't find\n");
         exit(1);
     }
     
@@ -295,9 +291,6 @@ int ap_unmount(const char *path)
     B_return(de);
 }
 
-/*任何一个线程调用close都会立即关闭描述符释放file结构
- *这样其它的线程用同一fd就访问不到此文件了
- */
 int ap_close(int fd)
 {
     if (fd < 0 || fd > _OPEN_MAX) {
@@ -319,9 +312,8 @@ int ap_close(int fd)
     pthread_mutex_unlock(&file->file_lock);
     pthread_mutex_unlock(&file_info.files_lock);
     
-    if (file->f_ops->release != NULL) {
+    if (file->f_ops != NULL && file->f_ops->release != NULL)
         file->f_ops->release(file, file->relate_i);
-    }
     AP_FILE_FREE(file);
     return 0;
 }
@@ -349,7 +341,7 @@ ssize_t ap_read(int fd, void *buf, size_t len)
     pthread_mutex_lock(&file->file_lock);
     pthread_mutex_unlock(&file_info.files_lock);
     
-    if (file->f_ops->read == NULL) {
+    if (file->f_ops == NULL ||  file->f_ops->read == NULL) {
         pthread_mutex_unlock(&file->file_lock);
         errno = ESRCH;
         return -1;
@@ -383,7 +375,8 @@ ssize_t ap_write(int fd, void *buf, size_t len)
     file = file_info.file_list[fd];
     pthread_mutex_lock(&file->file_lock);
     pthread_mutex_unlock(&file_info.files_lock);
-    if (file->f_ops->write == NULL) {
+    if (file->f_ops == NULL || file->f_ops->write == NULL) {
+        pthread_mutex_unlock(&file->file_lock);
         errno = ESRCH;
         return -1;
     }
@@ -404,7 +397,7 @@ off_t ap_lseek(int fd, off_t ops, int origin)
     struct ap_file *file;
     
     pthread_mutex_lock(&file_info.files_lock);
-    if (file_info.file_list[fd] == NULL) {
+    if (file->f_ops == NULL || file_info.file_list[fd] == NULL) {
         errno = EBADF;
         return -1;
     }
@@ -412,6 +405,7 @@ off_t ap_lseek(int fd, off_t ops, int origin)
     pthread_mutex_lock(&file->file_lock);
     pthread_mutex_unlock(&file_info.files_lock);
     if (file->f_ops->llseek == NULL) {
+        pthread_mutex_unlock(&file->file_lock);
         errno = ESRCH;
         return -1;
     }
@@ -439,7 +433,7 @@ int ap_mkdir(char *path, unsigned long mode)
     
     ap_fpthr = pthread_getspecific(file_thread_key);
     if (ap_fpthr == NULL) {
-        fprintf(stderr, "ap_thread didn't find\n");
+        ap_err("ap_thread didn't find\n");
         exit(1);
     }
     
@@ -475,9 +469,8 @@ int ap_mkdir(char *path, unsigned long mode)
     
     s_make = par_indic->cur_inode->i_ops->mkdir(par_indic); //由此函数直接负责将ap_inode插入到链表中
     AP_INODE_INICATOR_FREE(par_indic);
-    if (s_make == -1) {
+    if (s_make == -1)
         return -1;
-    }
     return 0;
 }
 
@@ -496,7 +489,7 @@ int ap_unlink(const char *path)
     
     ap_fpthr = pthread_getspecific(file_thread_key);
     if (ap_fpthr == NULL) {
-        fprintf(stderr, "ap_thread didn't find\n");
+        ap_err("ap_thread didn't find\n");
         exit(1);
     }
     
@@ -541,8 +534,8 @@ int ap_unlink(const char *path)
     
     int o = 0;
     
-    if (link == 0) {     //已经没有其它目录链接此文件
-        if (op_inode->parent->i_ops->unlink != NULL) {
+    if (link == 0) {   
+        if (op_inode->i_ops != NULL && op_inode->i_ops->unlink != NULL) {
             int unlik_s = op_inode->parent->i_ops->unlink(op_inode);
             if (unlik_s == -1) {
                 errno = EBUSY;
@@ -575,7 +568,7 @@ int ap_link(const char *l_path, const char *t_path)
     
     ap_fpthr = pthread_getspecific(file_thread_key);
     if (ap_fpthr == NULL) {
-        fprintf(stderr, "ap_thread didn't find\n");
+        ap_err("ap_thread didn't find\n");
         exit(1);
     }
     
@@ -665,7 +658,7 @@ int ap_rmdir(const char *path)
     
     ap_fpthr = pthread_getspecific(file_thread_key);
     if (ap_fpthr == NULL) {
-        fprintf(stderr, "ap_thread didn't find\n");
+        ap_err("ap_thread didn't find\n");
         exit(1);
     }
     
@@ -712,7 +705,7 @@ int ap_rmdir(const char *path)
         return -1;
     }
     pthread_mutex_unlock(&op_inode->inode_counter.counter_lock);
-    if (op_inode->i_ops->rmdir != NULL) {
+    if (op_inode->i_ops != NULL && op_inode->i_ops->rmdir != NULL) {
         rm_s = op_inode->i_ops->rmdir(final_indc);
         if (rm_s != 0) {
             pthread_mutex_unlock(&op_inode->ch_lock);
@@ -737,9 +730,9 @@ int ap_rmdir(const char *path)
     pthread_mutex_unlock(&op_inode->mount_inode->mt_ch_lock);
     op_inode = convert_to_real_ind(op_inode);
     
-    if (op_inode->mount_inode->real_inode == op_inode) {
+    if (op_inode->mount_inode->real_inode == op_inode)
         AP_INODE_FREE(op_inode->mount_inode);
-    }
+
     AP_INODE_INICATOR_FREE(final_indc);
     return 0;
 }
@@ -752,7 +745,7 @@ int ap_chdir(const char *path)
     }
     struct ap_file_pthread *ap_fpthr;
     ap_fpthr = pthread_getspecific(file_thread_key);
-    if (strcmp(path, AP_REWIND_DIR) == 0) {
+    if (strcmp(path, AP_REWIND_DIR_BEYOND_ROOT) == 0) {
         ap_inode_put(ap_fpthr->c_wd);
         ap_fpthr->c_wd = ap_fpthr->m_wd;
         return 0;
@@ -806,9 +799,9 @@ AP_DIR *ap_open_dir(const char *path)
     }
     
     int get = walk_path(indc);
-    if (get == -1) {
+    if (get == -1)
         B_return(NULL);
-    }
+
     if (!indc->cur_inode->is_dir) {
         errno = ENOTDIR;
         B_return(NULL);
@@ -860,9 +853,9 @@ int ap_closedir(AP_DIR *dir)
         errno = EINVAL;
         return -1;
     }
-    if (dir->dir_i->i_ops->closedir != NULL) {
+    if (dir->dir_i->i_ops->closedir != NULL)
         dir->dir_i->i_ops->closedir(dir->dir_i);
-    }
+
     AP_DIR_FREE(dir);
     return 0;
 }
